drobne zmiany, które znikły w wyniku returna

+++ chinski_sop/sop/memory/logical_memory.h
@@ -10,22 +10,22 @@ namespace memory
 {
 struct page //struktura strony
 {
-  int8_t frame_number;//nr ramki do jakej przynale¿y stron
+  uint16_t frame_number;//nr ramki do jakej przynale¿y stron
   char valid_invalid;//czy strona w pamiêci czy nie, 0 - invalid, 1 - valid
 };
 class LogicalMemory
 {
 public:	
-  explicit LogicalMemory(int8_t number_of_page);//konstruktor tworz¹cy tablicê stron o podanej liczbie stron
-  void setPageTableSize(int8_t size);//ustawia liczbê stron
+  explicit LogicalMemory(uint8_t number_of_page);//konstruktor tworz¹cy tablicê stron o podanej liczbie stron
+  void setPageTableSize(uint8_t size);//ustawia liczbê stron
   uint8_t getPageTableSize();//zwraca liczbê stron
-  page* getPage(int8_t page_nr);//zwraca wskazan¹ stronê
-  void setPage(int8_t frame_nr,char val_inv,int8_t page_nr);//ustawia poszczególnej strony i jej parametrów
+  page* getPage(uint8_t page_nr);//zwraca wskazan¹ stronê
+  void setPage(uint8_t frame_nr,char val_inv,uint8_t page_nr);//ustawia poszczególnej strony i jej parametrów
   
 protected:
 
 private:
-	int8_t page_table_size; //zmienna odpowiadaj¹ca za liczbê  stron
+	uint8_t page_table_size; //zmienna odpowiadaj¹ca za liczbê  stron
   boost::shared_ptr<page[]> page_table;//tabela stron, rekordów tyle ile ramek
 };
 //page_table = boost::shared_ptr<page[]>(new page[page_table_size]); w konruktorze zrobic
